+++ src/pancake/SeedHit.cpp
@@ -38,9 +38,10 @@ void CalcHitCoverage(const std::vector<SeedHit>& hits, int32_t hitsBegin, int32_
             }
             throw std::runtime_error(oss.str());
         }
-        coveredBasesQuery += std::min(hits[i].querySpan, (hits[i].queryPos - hits[i - 1].queryPos));
-        coveredBasesTarget +=
-            std::min(hits[i].targetSpan, (hits[i].targetPos - hits[i - 1].targetPos));
+        coveredBasesQuery += std::min(static_cast<int32_t>(hits[i].querySpan),
+                                      (hits[i].queryPos - hits[i - 1].queryPos));
+        coveredBasesTarget += std::min(static_cast<int32_t>(hits[i].targetSpan),
+                                       (hits[i].targetPos - hits[i - 1].targetPos));
     }
 }
